+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- quickfix publisher to always write an ndjson directly to S3
+  (will become outdated by MX-1808)
+
 ### Deprecated
 
 ### Removed
@@ -37,8 +40,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - switch out wikidata query by names with query by id
 - update seq-repo, odk, datscha-web, sumo, synopse test data
 - simplify tests by using settings fixture
-- quickfix publisher to always write an ndjson and upload it to S3
-  (will become outdated by MX-1808)
 
 ### Fixed